+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added `create_id` to `compas_ghpython.utilities`. (moved from `compas_fab`)
+* Added new predicates in `predicates_2` for points and polygons (XY) : `polygon_to_polygon_relationship_xy`, `point_to_polygon_relationship_xy`, `is_intersection_polygon_polygon_xy`
 
 ### Changed